pyproject.toml
[project]
name = "email-library"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.45.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.2",
    "uvicorn>=0.38.0",
]
